Bump version to 4.5.3
Be explicit about the builds and providers that we want to produce.
linux-ppc64le is (at the time of writing) a little temperamental, and
its default runs on Travis, which is phasing out its cheap open-source
support.

Setting linux, osx and win to azure just codifies the old default (note
that conda-smithy interpreted a conda-forge.yml file with these keys
missing as being implicitly "azure", not "default" which seems a little
strange).  linux-aarch64 and linux-ppc64le are set to "default" which
lets conda-smithy pick a suitable machine.  At the time of writing,
this defaults to Travis (since it has actual ppc46le hardware), while
Azure would work in theory, but the hardware was emulated.